requirements.txt
fastapi==0.114.0
psycopg2-binary==2.9.9
pydantic==2.9.1
python-dotenv==1.0.1
SQLAlchemy==2.0.34
uvicorn==0.30.6
alembic==1.13.2
asyncpg==0.29.0
beautifulsoup4==4.12.3
aiohttp==3.10.6
xmltodict==0.13.0
